requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cryptography==37.0.2
numpy==1.22.4
pyshp==2.3.0
Shapely==1.8.2
Fiona
geopy==2.2.0
pyproj==3.2.1
geopandas==0.10.2
Flask==2.1.1
geographiclib==1.52
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
munch==2.5.0
outcome==1.1.0
pandas==1.4.2
Pillow==9.1.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-geohash==0.8.5
pytz==2022.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
flask_cors
